adc triggers working, peak index scaling fixed for use in PID, yay!
It works! Yay!
To use the full range for the peakIndexes, set the peak time range equal to the acquisition duration .

Now, I want to check if I can modify the asg so that it stops when it finishes a curve period and then it waits for a new trigger