hk expanded, linearizer implemented, started adding ramp
Sorry, this commit is massive...
Anyway, now each digital pin output can be obtained from: memory, another pin (which can be different for each pin), a bit coming from the dsp module and the fastSwitch

linearizer added. Though I ran out of LUT blocks, so I had to disable the IIR block (which I am currently not using). Would be great to convince the compiler to use LUTRAM blocks instead... Oh, and I've added some proper triggers for the scope and dsp modules, so that you're not limited to using  pin 0P for all the external triggers (but I haven't implemented them in python yet)

a lot of modifications in the python code were necessary... I added a new register type that handles arrays of other registers. It could be useful to change some of the already existing arrays

for the ramp, the actual working ramp module and relative python code are missing, but the dsp level should already be working

Also, I've brought back the peak finder, that I erroneously updated to a newer version that has also a "height check" on the peaks (but for the moment it's not necessary, and I would need to heavily debug it). Next time...