fixed a glitch in the fpga
for some reason, writing the PID setpoint would write some bits in the peak ranges. Just to be sure, I made the write/read enablers (sys_wen and sys_ren) clocked registers, so there should be no way for 2 modules to have the enablers at the same time. Now it works (as far as I saw), but I'm not sure if the problem is fixed completely, or just that I got lucky with the routing of this firmware