now peak values will be saved in the config file
Finally, I've made it! Yay
So, the problem was that the values of the peaks/secondary pitayas were not saved because the file only saves the attributes contained in _setup_attributes. Then, the written values were not loaded because the loading is done at the end of the pyrpl __init__. Now I added a list of "submodules", where each submodule adds itself in (not very elegant, but I didn't find a better and faster option), so that pyrpl can load the data at the end of the initialization
Nice