Added acquisitions smaller than 1e-4. super fast peak detection!
This update was quite easy, I'm surprised. Also, no FPGA modifications required. Yay
Essentially, to reach higher scan frequencies with the AOM, and since the scope module already reaches the minimum sampling rate, the only way to increase the acquisition frequency is just to use the maximum frequency and just consider less samples. The peaks can be re-triggered before the real acquisition has ended, thus we can do many scan. The only limitation is that we're gonna have less samples to work on (as expectable). I separated the scope.duration property into 2, the actual acquisition duration and the ratio of data we actually keep.
With faster frequencies, I also needed to reduce the debounce delay of the scope (which fortunately was quite easy), and the peak ranges now are scaled when the "usedAcquisitionRatio" has changed. No biggie though

Also, I started implementing the automatic asg amplitude control with the second peak detection+PID